Migrate editing chain to ChunkCoord with sub-scale placement tests
Protocol SetVoxel uses [i64; 4] wire format (raw ChunkCoord bits).
All server/client logic uses [ChunkCoord; 4]: world_to_chunk_at_scale,
apply_voxel_edit, determine_chunk_scale, and the full edit chain.
Conversion happens at protocol boundaries only.

Test helpers (remove/place_block_along_ray) rewritten to use ChunkCoord
natively — no more i32 truncation for sub-voxel positions. Added 8 tests
covering scale -1 and -2 placement against various hit block scales,
including fractional grid snapping, all 4 axes, and negative coordinates.

+++ crates/polychora/src/shared/protocol.rs
@@ -119,7 +119,7 @@ pub enum ClientMessage {
         look: [f32; 4],
     },
     SetVoxel {
-        position: [i32; 4],
+        position: [i64; 4],
         block: BlockData,
     },
     SpawnEntity {
